Cart modal functionality 100%
