+++ src/components/CartModal/CartModal.scss
@@ -1,0 +1,180 @@
+$c-black: #1d1f22;
+$c-green: #5ece7b;
+
+.items-wrapper {
+  display: flex;
+  flex-direction: column;
+  row-gap: 40px;
+}
+
+.item {
+  display: flex;
+  column-gap: 48px;
+
+  &__details {
+    position: relative;
+  }
+
+  &__image {
+    width: 121px;
+    height: 190px;
+  }
+
+  &__preview-image {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+  }
+
+  &__title,
+  &__description {
+    font-weight: 300;
+    font-size: 16px;
+    line-height: 160%;
+  }
+
+  &__description {
+    margin-bottom: 4px;
+  }
+
+  &__price {
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 160%;
+  }
+
+  &__size-label,
+  &__color-label {
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 16px;
+  }
+
+  &__size-label,
+  &__size-button,
+  &__price,
+  &__color-label {
+    margin-bottom: 8px;
+  }
+
+  &__size-button {
+    background-color: #fff;
+    border: 1px solid $c-black;
+    font-weight: 400;
+    font-size: 14px;
+    line-height: 160%;
+    width: 24px;
+    height: 24px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    cursor: pointer;
+
+    &--selected {
+      background-color: $c-black;
+      color: #fff;
+    }
+  }
+
+  &__color-button {
+    width: 20px;
+    height: 20px;
+    cursor: pointer;
+
+    &--selected {
+      border: 1px solid $c-green;
+    }
+  }
+
+  &__size-buttons,
+  &__color-buttons {
+    display: flex;
+    column-gap: 8px;
+  }
+
+  &__quantity-button {
+    background-color: #fff;
+    border: 1px solid $c-black;
+    width: 24px;
+    height: 24px;
+    cursor: pointer;
+  }
+
+  &__quantity {
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 160%;
+
+    &-controls {
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      align-items: center;
+      position: absolute;
+      right: -40px;
+      top: 0;
+      bottom: 0;
+    }
+  }
+
+  &__total {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    font-size: 16px;    
+    margin-top: 41px;
+    margin-bottom: 32px;
+
+    &-label {
+      font-family: 'Roboto';
+      line-height: 18px;
+      font-weight: 500;
+    }
+
+    &-price {
+      font-weight: 700;
+      line-height: 160%;
+    }
+  }
+}
+
+.cart-modal {
+  position: absolute;
+  background-color: #fff;
+  top: 0;
+  right: 72px;
+  padding: 32px 16px;
+
+  &__title {
+    margin-bottom: 32px;
+  }
+
+  &-controls {
+    display: flex;
+    column-gap: 12px;
+  
+    &__view-bag-button {
+      font-weight: 600;
+      font-size: 14px;
+      line-height: 120%;
+      color: $c-black;
+      width: 140px;
+      height: 43px;
+      background: #fff;
+      border: 1px solid $c-black;
+      cursor: pointer;
+    }
+  
+    &__check-out-button {
+      font-weight: 600;
+      font-size: 14px;
+      line-height: 120%;
+      color: #fff;
+      width: 140px;
+      height: 43px;
+      background: $c-green;
+      cursor: pointer;
+    }
+  }
+}
+
