--- src/components/CartOverlay/Cart.scss
@@ -1,116 +1,0 @@
-$c-black: #1d1f22;
-$c-green: #5ece7b;
-
-.item {
-  &__title,
-  &__description {
-    font-weight: 300;
-    font-size: 16px;
-    line-height: 160%;
-  }
-
-  &__description {
-    margin-bottom: 4px;
-  }
-
-  &__price {
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 160%;
-  }
-
-  &__size-label,
-  &__color-label {
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 16px;
-  }
-
-  &__size-label,
-  &__size-button,
-  &__price,
-  &__color-label {
-    margin-bottom: 8px;
-  }
-
-  &__size-button {
-    background-color: #fff;
-    border: 1px solid $c-black;
-    font-weight: 400;
-    font-size: 14px;
-    line-height: 160%;
-    width: 24px;
-    height: 24px;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    cursor: pointer;
-
-    &--selected {
-      background-color: $c-black;
-      color: #fff;
-    }
-  }
-
-  &__color-button {
-    width: 20px;
-    height: 20px;
-    cursor: pointer;
-
-    &--selected {
-      border: 1px solid $c-green;
-    }
-  }
-
-  &__size-buttons,
-  &__color-buttons {
-    display: flex;
-    column-gap: 8px;
-  }
-
-  &__quantity-button {
-    width: 24px;
-    height: 24px;
-    cursor: pointer;
-  }
-
-  &__quantity {
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 160%;
-  }
-}
-
-.cart_overlay {
-  position: absolute;
-  background-color: #fff;
-  top: 0;
-  right: 72px;
-  padding: 32px 16px;
-}
-
-.cart-overlay-controls {
-  display: flex;
-  column-gap: 12px;
-
-  &__view-bag-button {
-    font-weight: 600;
-    font-size: 14px;
-    line-height: 120%;
-    color: $c-black;
-    width: 140px;
-    height: 43px;
-    background: #fff;
-    border: 1px solid $c-black;
-  }
-
-  &__check-out-button {
-    font-weight: 600;
-    font-size: 14px;
-    line-height: 120%;
-    color: #fff;
-    width: 140px;
-    height: 43px;
-    background: $c-green;
-  }
-}
