+++ src/components/Header/Header.scss
@@ -34,7 +34,7 @@
   cursor: pointer;
 }
 
-.cart {
+.cart-logo {
   position: relative;
 
   &::after,
